utils, download k_nearest csv
